Do not round position when dropping powerup / NPC

+++ src/npc/npc_bonus.cpp
@@ -91,7 +91,7 @@ void DropBonus(int A)
                 double ScreenTop = -vScreenY[1];
                 if(vScreen[1].Height > 600)
                     ScreenTop += vScreen[1].Height / 2 - 300;
-                int CenterX = -vScreenX[1] + vScreen[1].Width / 2;
+                double CenterX = -vScreenX[1] + vScreen[1].Width / 2;
 
                 if(g_config.hud_follows_player)
                 {
@@ -119,7 +119,7 @@ void DropBonus(int A)
                 double ScreenTop = -vScreenY[A];
                 if(vScreen[A].Height > 600)
                     ScreenTop += vScreen[A].Height / 2 - 300;
-                int CenterX = -vScreenX[A] + vScreen[A].Width / 2;
+                double CenterX = -vScreenX[A] + vScreen[A].Width / 2;
 
                 if(g_config.hud_follows_player)
                 {
